ts/features/bonus/siciliaVola/store/reducers/__test__/availableDestination.test.ts
import * as pot from "italia-ts-commons/lib/pot";
import { createStore } from "redux";
import { appReducer } from "../../../../../../store/reducers";
import { applicationChangeState } from "../../../../../../store/actions/application";
import {
  svGenerateVoucherAvailableDestination,
  svGenerateVoucherStart
} from "../../actions/voucherGeneration";
import { getTimeoutError } from "../../../../../../utils/errors";
import { AvailableDestinations } from "../../../types/SvVoucherRequest";

const genericError = getTimeoutError();
const mockResponse: AvailableDestinations = ["dest1", "dest2"];

describe("Test availableDestination reducer", () => {
  it("Initial state should be pot.none", () => {
    const globalState = appReducer(undefined, applicationChangeState("active"));
    expect(
      globalState.bonus.sv.voucherGeneration.availableDestinations
    ).toStrictEqual(pot.none);
  });
  it("Should be pot.none after if the voucher generation workunit starts", () => {
    const globalState = appReducer(undefined, applicationChangeState("active"));
    const store = createStore(appReducer, globalState as any);
    store.dispatch(svGenerateVoucherStart());

    expect(
      store.getState().bonus.sv.voucherGeneration.availableDestinations
    ).toStrictEqual(pot.none);
  });
  it("Should be pot.noneLoading after the first loading action dispatched", () => {
    const globalState = appReducer(undefined, applicationChangeState("active"));
    const store = createStore(appReducer, globalState as any);
    store.dispatch(svGenerateVoucherAvailableDestination.request());

    expect(
      store.getState().bonus.sv.voucherGeneration.availableDestinations
    ).toStrictEqual(pot.noneLoading);
  });
  it("Should be pot.some with the response, after the success action", () => {
    const globalState = appReducer(undefined, applicationChangeState("active"));
    const store = createStore(appReducer, globalState as any);
    store.dispatch(svGenerateVoucherAvailableDestination.success(mockResponse));

    expect(
      store.getState().bonus.sv.voucherGeneration.availableDestinations
    ).toStrictEqual(pot.some(mockResponse));
  });
  it("Should be pot.Error if is dispatched a failure after the first loading action dispatched", () => {
    const globalState = appReducer(undefined, applicationChangeState("active"));
    const store = createStore(appReducer, globalState as any);
    store.dispatch(svGenerateVoucherAvailableDestination.request());
    store.dispatch(svGenerateVoucherAvailableDestination.failure(genericError));

    expect(
      store.getState().bonus.sv.voucherGeneration.availableDestinations
    ).toStrictEqual(pot.toError(pot.none, genericError));
  });
});
